% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot state accidents for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number indicating for which state to plot accidents}

\item{year}{the year for which to plot accidents}
}
\value{
Plots the different accidents on a state map
for the given state and year
}
\description{
This function take a state (as a numeric value) and a year
and plots the accident that happens that year in that state.
}
\examples{
\dontrun{
fars_map_state(6, 2013)
}

}

