% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading a FARS csv data set}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the csv file from which FARS data should be read}
}
\value{
return a tbl_df with the FARS data of the input file
}
\description{
This function reads a csv file (provided as the argument "filename") and
returns the data as a \code{\link[dplyr]{tbl_df}} object. The intention
of the function is to be used to read in a FARS data set.
}
\examples{
\dontrun{
data1 <- fars_read("accident_2015.csv.bz2")
}

}

