% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years (as integers or characters)}
}
\value{
a list of tibble data frames, one for each of
the specified years. Each of the data frames contains
only the columns MONTH and year.
}
\description{
This function takes a vector of years and return a list of
data frames with month and years for the FARS data for the
input vector of years
}
\examples{
\dontrun{
datalist <- fars_read_years(list(2013, 2014))
}

}

