% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizing FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years (as integers or characters)}
}
\value{
a tibble data frame with a column for each
input year and an additional column "MONTH".
}
\description{
This function take a vector of years and returns a summary
of how many data point that are for each year and month
in the given range of years.
}
\examples{
\dontrun{
summary <- fars_summarize_years(list(2013, 2014))
}

}

