% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creating FARS data file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year for which to create the file name. Can be
either integer or character}
}
\value{
a character string with the file name of the FARS data
for the given year.
}
\description{
This function takes a year as input and returns the file name
of the file containing FARS data for that particular year.
}
\examples{
\dontrun{
filename1 <- make_filename(2015)
filename2 <- make_filename("2014")
}

}

